mise en place des restrictions d'acès suivant consignes ECF

+++ src/Controller/ProjectController.php
@@ -9,6 +9,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 #[Route('/project')]
 class ProjectController extends AbstractController
@@ -16,14 +17,40 @@ class ProjectController extends AbstractController
     #[Route('/', name: 'project_index', methods: ['GET'])]
     public function index(ProjectRepository $projectRepository): Response
     {
+        if (
+            !$this->isGranted('ROLE_ADMIN')
+            && !$this->isGranted('ROLE_TEACHER')
+            && !$this->isGranted('ROLE_STUDENT')
+            && !$this->isGranted('ROLE_CLIENT')
+        ) {
+            throw new AccessDeniedException();
+        }
+        if (
+            $this->isGranted('ROLE_STUDENT')
+            || $this->isGranted('ROLE_CLIENT')
+        ) {
+            $projects = $this->getUser()->getProjects();
+        } else {
+            $projects = $projectRepository->findAll();
+        }
+
+
         return $this->render('project/index.html.twig', [
-            'projects' => $projectRepository->findAll(),
+            'projects' => $projects,
+            'userName' => $this->getUser()->getLastname()
         ]);
     }
 
     #[Route('/new', name: 'project_new', methods: ['GET', 'POST'])]
     public function new(Request $request): Response
     {
+        if (
+            !$this->isGranted('ROLE_ADMIN')
+            && !$this->isGranted('ROLE_TEACHER')
+        ) {
+            throw new AccessDeniedException();
+        }
+
         $project = new Project();
         $form = $this->createForm(ProjectType::class, $project);
         $form->handleRequest($request);
@@ -45,6 +72,25 @@ class ProjectController extends AbstractController
     #[Route('/{id}', name: 'project_show', methods: ['GET'])]
     public function show(Project $project): Response
     {
+        if (
+            !$this->isGranted('ROLE_ADMIN')
+            && !$this->isGranted('ROLE_TEACHER')
+            && !$this->isGranted('ROLE_STUDENT')
+            && !$this->isGranted('ROLE_CLIENT')
+        ) {
+            throw new AccessDeniedException();
+        }
+
+        if (
+            $this->isGranted('ROLE_STUDENT')
+            || $this->isGranted('ROLE_CLIENT')
+        ) {
+            $projects = $this->getUser()->getProjects();
+
+            if (!$projects->contains($project)) {
+                throw new AccessDeniedException();
+            }
+        }
         return $this->render('project/show.html.twig', [
             'project' => $project,
         ]);
@@ -53,6 +99,20 @@ class ProjectController extends AbstractController
     #[Route('/{id}/edit', name: 'project_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Project $project): Response
     {
+        if (
+            !$this->isGranted('ROLE_ADMIN')
+            && !$this->isGranted('ROLE_TEACHER')
+            && !$this->isGranted('ROLE_STUDENT')
+        ) {
+            throw new AccessDeniedException();
+        }
+        if ($this->isGranted('ROLE_STUDENT')) {
+            $projects = $this->getUser()->getProjects();
+
+            if (!$projects->contains($project)) {
+                throw new AccessDeniedException();
+            }
+        }
         $form = $this->createForm(ProjectType::class, $project);
         $form->handleRequest($request);
 
@@ -71,7 +131,14 @@ class ProjectController extends AbstractController
     #[Route('/{id}', name: 'project_delete', methods: ['POST'])]
     public function delete(Request $request, Project $project): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$project->getId(), $request->request->get('_token'))) {
+        if (
+            !$this->isGranted('ROLE_ADMIN')
+            && !$this->isGranted('ROLE_TEACHER')
+        ) {
+            throw new AccessDeniedException();
+        }
+
+        if ($this->isCsrfTokenValid('delete' . $project->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($project);
             $entityManager->flush();
