+++ src/Controller/UserController.php
@@ -10,14 +10,21 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
-
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 #[Route('/user')]
 class UserController extends AbstractController
 {
+
     #[Route('/', name: 'user_index', methods: ['GET'])]
     public function index(UserRepository $userRepository): Response
     {
+        if (
+            !$this->isGranted('ROLE_ADMIN')
+        ) {
+            throw new AccessDeniedException();
+        }
+
         return $this->render('user/index.html.twig', [
             'users' => $userRepository->findAll(),
         ]);
@@ -26,6 +33,12 @@ class UserController extends AbstractController
     #[Route('/admin', name: 'user_index_admin', methods: ['GET'])]
     public function indexAdmin(UserRepository $userRepository): Response
     {
+        if (
+            !$this->isGranted('ROLE_ADMIN')
+        ) {
+            throw new AccessDeniedException();
+        }
+
         return $this->render('user/index.html.twig', [
             'users' => $userRepository->findByRole("ROLE_ADMIN"),
         ]);
@@ -34,6 +47,12 @@ class UserController extends AbstractController
     #[Route('/new', name: 'user_new', methods: ['GET', 'POST'])]
     public function new(Request $request, UserPasswordEncoderInterface $passwordEncoder): Response
     {
+        if (
+            !$this->isGranted('ROLE_ADMIN')
+        ) {
+            throw new AccessDeniedException();
+        }
+
         $user = new User();
 
 
@@ -61,19 +80,36 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'user_show', methods: ['GET'])]
     public function show(User $user): Response
     {
+        if (
+            !$this->isGranted('ROLE_ADMIN')
+        ) {
+            throw new AccessDeniedException();
+        }
+
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'user_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, User $user): Response
+    public function edit(Request $request, User $user, UserPasswordEncoderInterface $passwordEncoder): Response
     {
+        if (
+            !$this->isGranted('ROLE_ADMIN')
+        ) {
+            throw new AccessDeniedException();
+        }
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            $entityManager = $this->getDoctrine()->getManager();
+            $newpwd = $form->get('password')->getData();
+            $newEncodedPassword = $passwordEncoder->encodePassword($user, $newpwd);
+            $user->setPassword($newEncodedPassword);
+            $entityManager->persist($user);
+            $entityManager->flush();
 
             return $this->redirectToRoute('user_index');
         }
@@ -87,6 +123,12 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'user_delete', methods: ['POST'])]
     public function delete(Request $request, User $user): Response
     {
+        if (
+            !$this->isGranted('ROLE_ADMIN')
+        ) {
+            throw new AccessDeniedException();
+        }
+
         if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($user);
